% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{collect.LT2D.fits}
\alias{collect.LT2D.fits}
\title{Finds all LT2D fits produced by user}
\usage{
collect.LT2D.fits()
}
\description{
Creates a list where the entry at each index is
an LT2D.fit.function.object as produced by the \link{LT2D.fit}
function.
}
