% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfuns.R
\name{detect2DLT}
\alias{detect2DLT}
\title{2D line transect survey simulator}
\usage{
detect2DLT(x, hr, b, ystart, ny = 1000)
}
\arguments{
\item{x}{perpendicular distances at which detection probability is to be calculated}

\item{hr}{The hazard function (either a function or a character variable of the function name)}

\item{b}{parameter vector for hazard function \code{hr}.}

\item{ystart}{Max forward detectin distance, beyond which the detection hazard is 
assumed to be zero.}

\item{ny}{Number of equally-spaced points in the interval (0,\code{ystart}) to use to 
specify the hazard function. Defaults to 1000.}
}
\value{
A data frame with elements \code{$x} and  \code{$y} being the perpendicular and 
forward distances to detections.
}
\description{
Generates detections and associated forward distances (\code{y}) for a population
of objects with perpendicular distances given by \code{x}, using the hazard function 
specified by \code{h.fun}, with parameters specified by \code{b}, and maximum forward 
detection distance \code{ystart}.
}
\examples{
ystart = 1300
w = 1500
hr = "h1"
b = c(5,0.7)
pi.x = "pi.hnorm"
logphi = 6.6
L=1000
En = 100 # expected population size
dat = simpop2DLT(L,w,pi.x,logphi,En,fixed.n=TRUE) # fixed.n=TRUE fixes popn size a En
par(mfrow=c(1,2))
plot(dat,pch="+",main="population")
abline(v=0,lty=2)
hist(dat$x,breaks=seq(-w,w,length=21),xlab="Perp. distance",main="")
# now do LT2D survey of these points (only need their x-values for that)
seen = detect2DLT(dat$x,hr,b,ystart,ny=1000)
layout(matrix(c(1,1,2,3),nrow=2))
plot(seen,xlim=c(-w,w),pch="+",main=paste(nrow(seen),"Detectons"))
abline(v=0,lty=2)
hist(seen$x,breaks=seq(-w,w,length=21),xlab="Perp. distance",main="")
hist(seen$y,breaks=seq(0,ystart,length=21),xlab="Forward distance",main="")

}
