% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{dpdx}
\alias{dpdx}
\title{Numerical derivative of detection function}
\usage{
dpdx(h.fun, pars, ystart, w, nx = 101, ny = 101)
}
\arguments{
\item{h.fun}{Perpendicular distance distribution function}

\item{ystart}{maximum forward distance (y) at which objects could be detected}

\item{w}{Perpendicular distance truncation distance}

\item{nx}{number of x-values at which to evaluate the perpendicular distance detection function 
(which will give \code{(nx-1)}) numerical derivatives).}

\item{ny}{number of y-values at which to evaluate the hazard function.}
}
\value{
A vector of length \code{(nx-1)} containing the finite differences.
}
\description{
Calculates the numerical derivative of perpendicular distance detection function arising 
from a specified hazard function. Derivatives are approximated by simple finite differences.
h.fun,pars,ystart,w,nx=101,ny=101
}
\examples{
ystart = 1000
w=1000
hr.name = "ip0"; hr.pars=c(7, 0.25)
nx=101
deriv.x = dpdx(hr.name,hr.pars,ystart,w,nx=nx)
par(mfrow=c(1,2))
plotdetmodel(hr.name,hr.pars,ystart,xlim=c(0,w),what="px")
plot(seq(0,w,length=(nx-1)),deriv.x,type="l",xlab="x",ylab="Derivative")
abline(0,0,lty=2)

}
