% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{get.y.from.x}
\alias{get.y.from.x}
\title{y(x) for Sy(y(x)|x).}
\usage{
get.y.from.x(x, rmin)
}
\arguments{
\item{x}{numeric ; perpendicular distance of animal from the transect}

\item{rmin}{numeric ;  radius of the truncation zone. The distance below which
all perpendicular distances are rounded down to 0 - if this was done in the
the field.}
}
\value{
y ; forward distance
}
\description{
retrieves the forward distance y at which an animal meets the
circumference of the truncation zone, given its perpendicular distance x,
}
