% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTGoF.R
\name{gof.LT2D}
\alias{gof.LT2D}
\title{Goodness-of-fit for LT2D models.}
\usage{
gof.LT2D(fit, plot = FALSE)
}
\arguments{
\item{plot}{If TRUE, does Q-Q plot. Point corresponding to largest difference
between empirical and theoretical cdf (on which the Kolmogarov-Smirnov test is
based) is circled in red.}

\item{LT2D.fit.modek, }{as output by \code{\link{fit.yx}}}
}
\value{
list containing vectors 'x' and 'y', each containing their 2 p-values

list containing vectors 'x' and 'y', each containing their 2 p-values
}
\description{
Calculates goodness-of-fit in perpendicular and forward dimensions,
plots fit, and returns p-values. Returns two p-values for each dimension:
the Kolmogorov-Smirnov, and the Cramer von Mises.

Calculates goodness-of-fit in perpendicular and forward dimensions,
plots fit, and returns p-values. Returns two p-values for each dimension:
the Kolmogorov-Smirnov, and the Cramer von Mises.
}
\seealso{
\code{\link{fityx}} \code{\link{p.kolmogarov}}

\code{\link{fityx}} \code{\link{p.kolmogarov}}
}
