% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{hessianCalcs}
\alias{hessianCalcs}
\title{Adds covariance from Hessian}
\usage{
hessianCalcs(par, hessian, corrFlag = 0.8, output)
}
\arguments{
\item{par}{The fitted parameters}

\item{hessian}{- The hessian atrix produced by optim}

\item{output}{The object to which the variance covariance matrix should be added}

\item{corrFLag}{The float between 0 and 1 that determines above which threshold the 
correlation between two parameters is excessive}
}
\value{
The object \code{output} passed as an argument, with the covariance matrix added
}
\description{
Calculates covariance matrix from Hessian, does some checks, and adds it to the 
input object \code{output}.
}
