% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{invp1_replacement}
\alias{invp1_replacement}
\title{Adds inverse detection probability column to dataset}
\usage{
invp1_replacement(LT2D.df, LT2D.fit.obj)
}
\arguments{
\item{LT2D.df;}{an input data frame of the same format as required by NDest.}

\item{LT2D.fit}{output from \code{\link{fit.yx}}.}
}
\description{
Adds inverse detection probability column to dataset, so that the NDest function
has the input it requires to produce its estimations.
}
