% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jitterzeros}
\alias{jitterzeros}
\title{Calculates perp. dist. detection function approximation}
\usage{
jitterzeros(x, y, xcut, ycut, anground.degree, seed = NULL)
}
\arguments{
\item{x}{Observed perpendicular distance data}

\item{y}{Observed forward distance data}

\item{xcut}{Perpendicular distance within which rounding is assumed to have occurred}

\item{ycut}{Forward distance within which rounding is assumed to have occurred}

\item{anground.degree}{Max angle to jitter x- and y-data}

\item{seed}{Random number seed for reproducibility. If NULL, random seed is generated}
}
\value{
Data frame with elements \code{$x} and \code{$y} being the new (jittered) 
perpendicular and forward distance observations, respectively.
}
\description{
Jitters the perpendicular distances (\code{x}s) within \code{xcut} of 
zero and the forward distances (\code{y}s) within \code{ycut} of 0. Does this by 
adding a random angle from a uniform random variable on (0,\code{anground.degree}) 
degrees to the angle of the \code{x}s within \code{xcut} of zero and subtracting 
a uniform random variable on (0,\code{anground.degree}) degrees from the angle of 
the \code{y}s within \code{ycut} of zero.
}
\examples{
xs = seq(0,w,length=100)
ys = seq(0,ystart,length=100)
h.fun.name = "h1"
b=c(-7.3287948, 0.9945317)
h.fun = match.fun(h.fun.name) # make h.fun the function specified via a character variable
p.vals = p.approx(ys,xs,h.fun,b) # detection function values to plot
plot(ys,p.vals,type='l',xlab='Perp. distance, x',ylab=expression(p(x)))

}
