% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{LT2D.fit}
\alias{LT2D.fit}
\title{Top level fitting and abundance estimation function for LT2D user}
\usage{
LT2D.fit(
  DataFrameInput,
  hr,
  b,
  ystart,
  pi.x,
  logphi,
  w,
  formulas = NULL,
  ipars = NULL,
  xpars = NULL,
  ypars = NULL,
  rmin = 0,
  control = list(),
  hessian = FALSE,
  corrFlag = 0.999,
  debug = FALSE
)
}
\arguments{
\item{DataFrameInput, }{data.frame ; data frame with required columns
 stratum (stratum number of observation), transect (transect number of
observation, object(object number of detection, NA if transect with no
detections), size (group size of detection), area (stratum area of observation)
and L (transect length of transect).}

\item{hr}{character; name of hazard rate to fit}

\item{b}{numeric; vector of start parameters for hr}

\item{ystart}{numeric; furthest possible forwards distance at which we can
detect animals.}

\item{pi.x}{character; name of perpendicular density function to fit}

\item{logphi}{numeric; vector of start parameters for pi.x}

\item{w}{numeric; perpendicular truncation distance}

\item{formulas}{parameter passed to \code{fityx} (specification of covariate models, maybe?)}

\item{ipars}{parameter passed to \code{fityx} (parameters of covariate model for intercept, maybe?)}

\item{xpars}{parameter passed to \code{fityx} (parameters of covariate model in x-direction, maybe?)}

\item{ypars}{parameter passed to \code{fityx} (parameters of covariate model in y-direction, maybe?)}

\item{rmin}{numeric; radial distance below which all values were rounded down to 0}

\item{control}{list passed to \code{optim} to control optimasation (see \code{optim} help)}

\item{hessian}{boolean indicating whether or not Hessian matrix is to be returned}

\item{corrFlag}{numeric; value above which correlation flag is raised}

\item{debug}{prints some stuff to help debugging if TRUE}
}
\value{
An object of class \code{LT2D.fit.function.object}, which is a list containing three objects:
\itemize{
 \item{invp} {A data frame with one row per detection and these columns:
  \itemize{
   \item{x} {The x-coordinate of the detection.}
   \item{y} {The y-coordinate of the detection.}
   \item{stratum} {The stratum in which the detection occurred.}
   \item{transect} {The transect on which the detection occurred.}
   \item{L} {The length of the transect.}
   \item{area} {The area of the stratum.}
   \item{object} {The object ID number.}
   \item{size} {The object group size.}
   \item{invp} {The inverse of the estimated detection probability of the object.}
   }
  }
 \item{ests} {A data frame in which each row contains the density and abundance estimates 
 for a stratum (with last row being for the total across all strata), with these columns:
  \itemize{
   \item{stratum} {The stratum.}
   \item{n} {The number of detections in the stratum.}
   \item{k} {The number of transects in the stratum.}
   \item{L} {The total transect length in the stratum.}
   \item{covered.area} {The covered area (2WL) in the stratum.}
   \item{stratum.Area} {The area of the stratum.}
   \item{Dgroups} {The estimated density of groups.}
   \item{Ngroups} {The estimated abundance of groups.}
   \item{mean.size} {The estimated mean group size.}
   \item{D} {The estimated density of individuals.}
   \item{N} {The estimated abundance of individuals.}
   }
  }
 \item{fit} {The output from the core fitting function \code{fityx}. This has a bunch of 
 input data reflection and input parameter reflection (not detailed here) as well as the following 
 estimates
 \itemize{
  \item{CVpar} {Estimated coefficients of variation of each model parameter.}
  \item{corr} {Parameter estimate correlation matrix.}
  \item{vcov} {Parameter estimate covariance matrix.}
  \item{par} {Parameter estimates.}
  \item{value} {Value of the negative log-likelihood at the MLE.}
  \item{AIC} {Aaike's information criterion value.}
  }
 }
}
}
\description{
This function is a wrapper for the fityx function. It takes
in a 'Distance' style data frame, type-checks it, finds the MLEs using
\code{\link{fityx}} and then calls an abundance estimation function \code{\link{NDest}}
to obtain density and abundance estimates by stratum.
}
