% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{LT2D.FitObjectMaker}
\alias{LT2D.FitObjectMaker}
\title{LT2D fit object maker}
\usage{
LT2D.FitObjectMaker(
  par,
  value,
  counts,
  convergence,
  message,
  hessian,
  error,
  hr,
  pi.x,
  ystart,
  w,
  b,
  logphi,
  AICval,
  dat,
  MLEvcov,
  CVpar,
  corr,
  p0,
  rmin
)
}
\arguments{
\item{par}{numeric scalar or vector; values of fitted MLE parameters.}

\item{value}{numeric; value of the negative log likelihood evaluated at the MLEs.}

\item{counts}{integer, two-element vector which represents the number of
likelihood (and gradient function) evaluations performed by the optimisation
routine (\code{\link{optim}})}

\item{convergence}{integer; optim's system for describing fit behaviour. see
\code{\link{optim}}}

\item{message}{NULL or character; optim's output message, if any.}

\item{hessian}{matrix or NULL; solved hessian matrix of MLEs}

\item{error}{logical; boolean; has an error occurred or were any flags raised?}

\item{hr}{character; name of the user-chosen hazard function}

\item{pi.x}{character; name of the user-chosen density}

\item{ystart}{numeric; value of user-chosen ystart value}

\item{w}{numeric; user-chosen perpendicular truncation distance}

\item{b}{numeric; vector of user-chosen start values for hazard rate MLE}

\item{logphi}{numeric; scalar or vector of user-chosen start values
for density rate MLE}

\item{AICval}{numeric; Akaike's Information Criterion of the model (uncorrected)}

\item{dat}{data.frame; data set used to fit the model.}

\item{MLEvcov}{matrix; variance covariance matrix of MLEs}

\item{CVpar}{numeric vector; Coefficient of variation of the MLEs}

\item{corr}{matrix; matrix of correlations for the MLEs}

\item{p0}{numeric; estimated probability of detection at distance (0,0)}
}
\value{
LT2Dfit object
}
\description{
Used by \code{\link{fityx}} to produce a custom object which
contains relevant information pertaining to the produced fit, and which could
be needed by fitting and GOF functions. It type-checks all of the values to
ensure the user is never presented with an object with unexpected content.
}
\seealso{
\code{\link{fityx}}
}
