% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT2D.R
\docType{package}
\name{LT2D}
\alias{LT2D}
\alias{LT2D-package}
\title{LT2D: A package implementing 2-dimensional line transect sampling}
\description{
Functions for data generation and fitting of 2-dimensional line transect sampling
}
\references{
Borchers, D.L. and Cox, K.J. 2017 Distance sampling detection functions: 2D or not 2D?. 
Biometrics 73: 593-602..
}
\author{
Cal Fagard-Jenkin, David Borchers
}
\keyword{2D}
\keyword{line}
\keyword{transect}
