% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{NDest}
\alias{NDest}
\title{Estimates density and abundance.}
\usage{
NDest(dat, LT2D.fit)
}
\arguments{
\item{dat}{DS data frame. Must have cols "stratum","area","transect","L","size","object","x","y"
(and possibly others).}

\item{LT2d.fit}{Object of class \code{LT2D.model.fit}, normally obtained as output from \code{\link{fit.hmltm}}.}
}
\description{
Horvitz-Thompson like estimation of density and abundance of groups and of individuals, as well as
of group size (estimated as the ratio of individual density and group density estimates). Produces
estimates by stratum and over all strata.
}
