% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTGoF.R
\name{p.kolmogarov}
\alias{p.kolmogarov}
\title{Kolmogarov-Smirnov goodness-of-fit p-value.}
\usage{
p.kolmogarov(x, inf = 1000, dp = 1e-04)
}
\arguments{
\item{x}{value of Kolmogarov-distributed random variable at which to evaluate.}

\item{inf}{approximation to infinity (a large number).}

\item{dp}{approximation convergence criterion.}
}
\value{
1 - p-value for Kolmogarov distribution
}
\description{
Kolmogarov-Smirnov goodness-of-fit p-value calculation.
}
\details{
Calculates p-value for Kolmogarov distribution at x, approximating infite sum
\code{sqrt(2*pi)/x*sum{i=1}^infty exp(-(2i-1)^2*pi^2/(8x^2)))}
by a finite sum to inf (default 1000) if sum to inf and inf+1 differ by less
than dp (default 1e-4), else sum until difference is less than dp.
}
