% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{pi.chnorm.i.exp.term}
\alias{pi.chnorm.i.exp.term}
\title{Helper function for pi.chnorm.i}
\usage{
pi.chnorm.i.exp.term(x, logphi)
}
\arguments{
\item{x}{numeric ; perpendicular distance x}

\item{logphi}{; numeric vector of length two of parameters}
}
\description{
function which evaluates part of the
expression of its namesake function. Useful for
allowing numeric integration of this expression
}
