% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{pi.chnorm.i}
\alias{pi.chnorm.i}
\title{Half normal density allowing for non-zero intercept}
\usage{
pi.chnorm.i(x, logphi, w)
}
\arguments{
\item{x}{prependicular trackline distance}

\item{logphi}{; numeric vector of length two}

\item{w}{perpendicular truncation distance}
}
\value{
numeric scalar or vector
}
\description{
Half normal perpendicular density function with
the ability to accomodate for positive density at x=0.
}
