% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{pi.x.vector.input}
\alias{pi.x.vector.input}
\title{density function with vecotr inputs}
\usage{
pi.x.vector.input(xVector, pi.x.Function, logphi, w)
}
\arguments{
\item{xVector}{numeric vector; perpendicular distances of animals from the transect}

\item{pi.x.Function}{function ; the density function we wish to call}

\item{logphi}{numeric vector; parameters for the density function}

\item{w}{numeric ; perpendicular truncation distance}
}
\value{
vector ; pi.x evaluated at each of the values in xVector
}
\description{
The integrate function requires the ability to call
function it integrates over with a vector of values. The
density distribution functions do not allow this, this function
acts as a wrapper, accepting a vector of x value inputs.'
}
