% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{plot.LT2D.fit.function.object}
\alias{plot.LT2D.fit.function.object}
\title{Plot LT2D fit}
\usage{
\method{plot}{LT2D.fit.function.object}(fit, ...)
}
\arguments{
\item{fit}{fitted object output by \code{\link{LT2D.fit}}}

\item{..., }{parameters to be passed to \code{\link{plot.LT2D.fit.object}}}
}
\description{
acts as a wrapper for the appropriate
plotting function.
}
\seealso{
\code{\link{LT2D.fit}}
}
