% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{plot.LT2D.fit.object}
\alias{plot.LT2D.fit.object}
\title{Plot fitted curves of an LT2D.fit.object}
\usage{
\method{plot}{LT2D.fit.object}(
  fit,
  xbins = 20,
  ybins = 20,
  smooth.fy = FALSE,
  addrug = FALSE,
  covar.row = FALSE
)
}
\arguments{
\item{fit}{fitted object output by \code{\link{fityx}}}

\item{xbins}{int;number of histogram bins to use for perpendicular distance}

\item{ybins}{int;number of histogram bins to use for forward distance}
}
\value{
invisibly returns the outputs of the three functions to which it
delagates tasks
}
\description{
Produces plots of fitted functions, pi, p and f for perpendicular
 distance x, as well as a plot of fitted fy for forward distance y. Calls
 \link{\code{plotfit.y}}, \link{\code{plotfit.x}} and
 \link{\code{plotfit.smothfy}} to produce the figures..
}
\seealso{
\code{\link{fityx}}
}
