% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{plotdetmodel}
\alias{plotdetmodel}
\title{Plot hazard and/or detection function model}
\usage{
plotdetmodel(h, pars, ystart, xlim, what = "px", nx = 101, ny = 101)
}
\arguments{
\item{pars}{parameters of hazard function}

\item{ystart}{maximum forward distance (y) at which objects could be detected}

\item{xlim}{range of perpendicular distances (x) to plot, comprising a vector of length 2: (xmin, xmax)}

\item{what}{Determines what is plotted: 'fxy' results in the pdf of forward detection distances, y, 
for the given x; 'pxy' results in the probability of detection BY each y being plotted; 
'px' results in the probability of detection at all, being plotted (over the whole specified perpendicular
distance range). Anything else results in all three of these being plotted.}

\item{nx}{number of x-values at which to evaluate the hazard function (ignored if xmin=xmax)}

\item{ny}{number of y-values at which to evaluate the hazard function}

\item{h.fun}{Detection hazard function}
}
\value{
If \code{what} is 'fxy', a matrix containing the pdf of y, for each x (matrix row) is 
returned. If 'pxy' a matrix with the probability of detection by each y for each x (matrix row) 
is returned. If 'px' a vector comprising the probability of detection at all for each x is 
returned. Anything else results in a list with all three of these being returned, in elements 
\code{$fmat}, \code{$pmat}, and \code{p}, respectively.
}
\description{
Depending on what is specified in the argument \code{what}, plots hazard and/or detection 
function model. Invisibly returns data to make plots.
}
\examples{
ystart = 2
h.fun = "h1"
pars=c(-2, 1)
xlim = c(1/10000,1)
plot.detmodel(h.fun,pars,ystart,xlim,what="all")
plot.detmodel(h.fun,pars,ystart,xlim,what="fxy")
plot.detmodel(h.fun,pars,ystart,xlim,what="pxy")
plot.detmodel(h.fun,pars,ystart,xlim,what="px")

}
