% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{plotpimodel}
\alias{plotpimodel}
\title{Plot perpendicular distance distribution model}
\usage{
plotpimodel(pi.fun, logphi, w, nx = 101)
}
\arguments{
\item{pi.fun}{Perpendicular distance distribution function}

\item{logphi}{Log of the parameters of perpendicular distance distribution function}

\item{w}{Perpendicular distance truncation distance}

\item{nx}{number of x-values at which to evaluate the function}

\item{ystart}{maximum forward distance (y) at which objects could be detected}
}
\value{
If \code{what} is 'fxy', a matrix containing the pdf of y, for each x (matrix row) is 
returned. If 'pxy' a matrix with the probability of detection by each y for each x (matrix row) 
is returned. If 'px' a vector comprising the probability of detection at all for each x is 
returned. Anything else results in a list with all three of these being returned, in elements 
\code{$fmat}, \code{$pmat}, and \code{p}, respectively.
}
\description{
Plot perpendicular distance distribution model. Invisibly returns data to make the plot.
}
\examples{
pi.fun.name = 'pi.hnorm'
logphi <- log(c(700))
w=1000
plotpimodel(pi.fun.name,logphi,w)

}
