% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{round.lik.denom}
\alias{round.lik.denom}
\title{denominator for rounded data likelihood}
\usage{
\method{round}{lik.denom}(pi.x, logphi, rmin, w)
}
\arguments{
\item{pi.x}{character ; user-chosen density model}

\item{logphi}{numeric vector ; parameters for pi.x}

\item{rmin}{numeric ; value below which radial distances have been
rounded to 0}
}
\value{
numeric
}
\description{
denominator for rounded data likelihood
}
