% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{round.lik.num}
\alias{round.lik.num}
\title{numerator for rounded data likelihood}
\usage{
\method{round}{lik.num}(pi.x, logphi, rmin, ymax, hr, b, w)
}
\arguments{
\item{pi.x}{character ; user-chosen density model}

\item{logphi}{numeric vector ; parameters for pi.x}

\item{rmin}{numeric ; value below which radial distances have been
rounded to 0}

\item{ymax}{numeric ; maximum forward distance at which we can detect animals}

\item{hr}{character ; user-chosen hazard rate}

\item{b}{numeric vector ; parameters for hr}

\item{w}{numeric ; perpendicular truncation distance}
}
\value{
numeric
}
\description{
numerator for rounded data likelihood
}
