% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{round.survival.prod.pi}
\alias{round.survival.prod.pi}
\title{intermediary function for rounded likelihood}
\usage{
\method{round}{survival.prod.pi}(x, pi.x, logphi, rmin, ymax, hr, b, w)
}
\arguments{
\item{x}{numeric ; perpendicular distance of animal from the transect}

\item{pi.x}{character ; user-chosen density model}

\item{logphi}{numeric vector ; parameters for pi.x}

\item{rmin}{numeric ; value below which radial distances have been
rounded to 0}

\item{ymax}{numeric ; maximum forward distance at which we can detect animals}

\item{hr}{character ; user-chosen hazard rate}

\item{b}{numeric vector ; parameters for hr}

\item{w}{numeric ; perpendicular truncation distance}
}
\value{
numeric
}
\description{
the function which we integrate to obtain the numerator
of the likelihood for rounded data
}
