% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfuns.R
\name{simpop2DLT}
\alias{simpop2DLT}
\title{Simulate population for 2D line transect survey}
\usage{
simpop2DLT(L, w, pi.x, logphi, En, fixed.n = FALSE)
}
\arguments{
\item{L}{Line length (length of strip)}

\item{w}{Strip half-width}

\item{pi.x}{Perpendicular distance distribution function (character or function)}

\item{logphi}{Parameters of \code{pi.x}.}

\item{En}{Expected number of objects to simulate (or exact number if \code{fixed.n} is TRUE)}

\item{fixed.n}{If TRUE exactly \code{En} objects are simulated, else Poisson number
of objects is simulated, with expected valuue \code{En}.}
}
\value{
A data frame with elements \code{$x} and  \code{$y} being the perpendicular and 
along-transect distances of objects.
}
\description{
Generates locations of a population with perpendicular distance
distribution specified by \code{pi.x}, with parameters specified by \code{logphi}, 
in a strip of length \code{L} and width \code{2w}. The number of individuals in 
the population is a Poisson random variable with mean \code{En}, unless \code{fixed.n}
is TRUE, in which case exactly \code{En} individuals are simulated.
}
\examples{
ystart = 1300
w = 1500
hr = "h1"
b = c(5,0.7)
pi.x = "pi.hnorm"
logphi = 6.6
L=1000
En = 100 # expected population size
dat = simpop2DLT(L,w,pi.x,logphi,En,fixed.n=TRUE) # fixed.n=TRUE fixes popn size a En
par(mfrow=c(1,2))
plot(dat,pch="+",main=paste("Population (N=",nrow(dat),")"))
abline(v=0,lty=2)
hist(dat$x,breaks=seq(-w,w,length=21),xlab="Perp. distance",main="")

}
