% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{simpson}
\alias{simpson}
\title{Integration using Simpson's 1/3 rule}
\usage{
simpson(f, lower, upper, ..., subdivisions = 50)
}
\arguments{
\item{lower}{Forward distance integration points}

\item{upper}{perpendicular distances at which detection probability is to be calculated}

\item{...}{Other arguments required by \code{fun}.}

\item{fun}{The  function to integrate (either a function or a character variable of the function name)}

\item{b}{parameter vector for function \code{fun}.}
}
\description{
Uses Simpsons 1/3 rule to integrate the specified hazard function \code{} from
\code{ystart} to 0, at every x-value in \code{xs}, using the y-values in \code{ys} as the 
integration points. Then calculates detection probability at each x as 1 - exp(-h.int), 
where h.int is this integral at each of the given x-values.
}
\examples{
ystart = 1300
hr = "h1"
b = c(5,0.7)
simpson(hr,0,ystart,x=0,b=b)
simpson(hr,0,ystart,x=10,b=b)

}
