% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{summarise.LT2D.models}
\alias{summarise.LT2D.models}
\title{Summarises all LT2D fits}
\usage{
summarise.LT2D.models()
}
\description{
Creates a data.frame with useful summary statistics
for all user fitted models produced by calling \link{LT2D.fit}
by using \link{collect.LT2D.fits} and then
\link{summary.collected.LT2D.models}.
}
