% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{summary.collected.LT2D.models}
\alias{summary.collected.LT2D.models}
\title{Creates a summary for a collected.LT2D.models object}
\usage{
\method{summary}{collected.LT2D.models}(objects)
}
\value{
data.frame containing:
Name ; names of the fitted model
AIC ; AIC of the fitted model
N ; Estimated Abundance from fitted model
D ; Estimated Density from fitted model
GoF ; A count, detailing how many of the four GoF tests
had a p value smaller than 0.05
}
\description{
Creates a data.frame with useful summary statistics
for all user fitted models as columns
}
