% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{Sy}
\alias{Sy}
\title{Calculates survivor function}
\usage{
Sy(x, y, ymax, b, hr, notInf = 10000)
}
\arguments{
\item{x}{perpendicular distance (scalar)}

\item{y}{forward distance  (scalar)}

\item{ymax}{largest forward distance}

\item{b}{vector of detection hazard function parameters}

\item{notInf}{Value of hfun evaluation above which the hazard is treated as
infinite when integrating, and therefore survival is assumed to be zero.}

\item{hfun}{detection hazard function}
}
\value{
Probability of making it from forward distance
\code{ymax} to \code{y} without being detected.
}
\description{
Calculates survivor function to forward distance \code{y},
for given perpendicular distance \code{x} and given forward distance range.
}
\details{
Calculates probability of making it from forward distance
\code{ymax} to \code{y} without being detected.
}
\examples{
\dontrun{
ymax=5
Sy(0,0.1,ymax,b,h2)
}
}
