% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DLTfunctions mixture.r
\name{p.approx}
\alias{p.approx}
\title{Calculates perp. dist. detection function approximation}
\usage{
p.approx(y, x, h.fun, b)
}
\arguments{
\item{y}{Forward distance integration points}

\item{x}{perpendicular distances at which detection probability is to be calculated}

\item{h.fun}{The hazard function (either a function or a character variable of the function name)}

\item{b}{parameter vector for hazard function \code{h.fun}.}
}
\value{
probability of detection at each of the \code{xs}.
}
\description{
Uses Simpsons 1/3 rule to integrate the specified hazard function \code{} from
\code{ystart} to 0, at every x-value in \code{xs}, using the y-values in \code{ys} as the 
integration points. Then calculates detection probability at each x as 1 - exp(-h.int), 
where h.int is this integral at each of the given x-values.
}
\examples{
xs = seq(0,w,length=100)
ys = seq(0,ystart,length=100)
h.fun.name = "h1"
b=c(-7.3287948, 0.9945317)
h.fun = match.fun(h.fun.name) # make h.fun the function specified via a character variable
p.vals = p.approx(ys,xs,h.fun,b) # detection function values to plot
plot(ys,p.vals,type='l',xlab='Perp. distance, x',ylab=expression(p(x)))

}
